ripped out react-native-router-flux and use expo router
