+++ app.json
@@ -2,6 +2,7 @@
   "expo": {
     "name": "Scout Skills",
     "slug": "scout-skills",
+    "scheme": "scout-skills",
     "privacy": "public",
     "platforms": [
       "ios",
@@ -36,7 +37,16 @@
     "androidStatusBar": {
       "backgroundColor": "#46BDFF",
       "translucent": false,
-      "barStyle":"light-content"
-    }
+      "barStyle": "light-content"
+    },
+    "plugins": [
+      "expo-router",
+      [
+        "expo-font",
+        {
+          "fonts": ["./assets/fonts/usuzi.ttf"]
+        }
+      ]
+    ]
   }
 }
