+++ app/index.js
@@ -1,0 +1,40 @@
+import React, { Component, useEffect, useState } from "react";
+import { StyleSheet, TouchableOpacity, Text, View,Platform } from "react-native";
+import * as Constants from "../component/Constants";
+import SkillSelectButton from "../component/SkillSelectButton";
+import { useFonts } from 'expo-font';
+
+export default function Home() {
+  const [fontsLoaded] = useFonts({
+    'usuzi': require('../assets/fonts/usuzi.ttf'),
+  });
+
+  if(!fontsLoaded) {
+    return null
+  }
+
+  return (
+    <View style={styles.container}>
+      <SkillSelectButton skill={Constants.CAMPING} textColor="#009F54" />
+      <SkillSelectButton skill={Constants.BACKWOODS} textColor="#559632" />
+      <SkillSelectButton skill={Constants.PIONEERING} textColor="#004E50" />
+
+      <SkillSelectButton skill={Constants.EMERGENCIES} textColor="#F57A41" />
+      <SkillSelectButton skill={Constants.HIKING} textColor="#F57921" />
+      <SkillSelectButton skill={Constants.AIR} textColor="#007CC2" />
+
+      <SkillSelectButton skill={Constants.PADDLING} textColor="#0668B3" />
+      <SkillSelectButton skill={Constants.ROWING} textColor="#0668B3" />
+      <SkillSelectButton skill={Constants.SAILING} textColor="#0060AA" />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 10,
+    flex: 1,
+    backgroundColor: "white",
+    paddingBottom: Platform.OS === 'ios' ? 20 : 10
+  },
+});
