+++ component/SkillSelectButton.js
@@ -1,20 +1,15 @@
 import React, { Component } from 'react';
 import { StyleSheet, Text, View, TouchableOpacity, Dimensions } from 'react-native';
-// import { Actions } from 'react-native-router-flux';
-
+import { Link } from "expo-router";
 export default class extends Component {
 
-  goToSkill() {
-    const { skill, textColor } = this.props
-    // Actions.skillScreen({ skill: skill, textColor: textColor })
-  }
 
   render() {
     const { skill, textColor } = this.props
     return (
-      <TouchableOpacity style={[styles.container, { backgroundColor: textColor }]} onPress={() => this.goToSkill()}>
-        <Text style={styles.text}>{this.props.skill}</Text>
-      </TouchableOpacity>
+      <Link style={[styles.container, { backgroundColor: textColor }]} href={`/skills/${skill}`}>
+        {this.props.skill}
+      </Link>
     );
   }
 }
@@ -27,9 +22,7 @@ const styles = StyleSheet.create({
     borderWidth: 3,
     justifyContent: 'center',
     alignItems: 'center',
-    marginTop: 5
-  },
-  text: {
+    marginTop: 5,
     color: 'white',
     fontFamily: 'usuzi',
     fontSize: 20
