+++ navigation/Routes.js
@@ -1,13 +1,7 @@
 import React from 'react'
 import { StyleSheet, TouchableOpacity, Text } from 'react-native';
-// import { Router, Scene,Actions } from 'react-native-router-flux'
 import Home from '../screens/Home'
-import { SettingsScreen } from '../screens/SettingsScreen';
-import { SkillScreen } from '../screens/SkillScreen'
-import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
-import { faCog } from '@fortawesome/free-solid-svg-icons';
 import * as Constants from '../component/Constants';
-import AboutApp from '../screens/AboutApp';
 
 const Routes = () => (
   <>
