+++ package.json
@@ -1,5 +1,5 @@
 {
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
@@ -18,9 +18,12 @@
     "@react-native-community/async-storage": "~1.12.0",
     "@react-native-community/checkbox": "^0.2.2",
     "expo": "~50.0.14",
+    "expo-constants": "~15.4.6",
     "expo-firebase-analytics": "8.0.0",
+    "expo-linking": "~6.2.2",
     "expo-mail-composer": "12.7.1",
-    "expo-status-bar": "1.11.1",
+    "expo-router": "~3.4.9",
+    "expo-status-bar": "~1.11.1",
     "firebase": "^7.23.0",
     "native-base": "^2.13.8",
     "react": "18.2.0",
@@ -30,13 +33,12 @@
     "react-native-gesture-handler": "^2.16.0",
     "react-native-reanimated": "^3.9.0",
     "react-native-responsive-fontsize": "^0.5.1",
-    "react-native-router-flux": "^4.3.1",
     "react-native-safe-area-context": "4.8.2",
-    "react-native-screens": "^3.31.1",
+    "react-native-screens": "~3.29.0",
     "react-native-svg": "^15.2.0"
   },
   "devDependencies": {
-    "babel-preset-expo": "^8.3.0"
+    "babel-preset-expo": "10.0.2"
   },
   "private": true
 }
